Fixed bugs in Allisons code. First two algorithms work with .txt conversion using additions to characterProcess and blockBrightness. CharacterSet still being developed. It is used to find the density of each character.

+++ src/TestMain.java
@@ -26,7 +26,7 @@ public class TestMain {
 		// Font myFont = new Font("Courier", Font.PLAIN, 12);
 		
 		BlockBrightness bb = new BlockBrightness();
-		double[][] dmatrix = bb.convertImageToBrightnessMatrix(bb.getImage());
+		int[][] dmatrix = bb.convertImageToBrightnessMatrix(bb.getImage());
 		CharacterProcess cp = new CharacterProcess();
 		cp.convertImageToAscii(dmatrix);
 //		double[][] test = cp.convertImageToBrightnessMatrix(cp.getImage());
